+++ src/mapear/modulos/mapear/aplicacion/__init__.py
@@ -2,6 +2,6 @@ from pydispatch import dispatcher
 
 from .handlers import HandlerParquetIntegracion
 
-from mapear.modulos.procesamiento.dominio.eventos import DatasetCreado
+from mapear.modulos.mapear.dominio.eventos import DatasetCreado
 
 dispatcher.connect(HandlerParquetIntegracion.handle_parquet_creado, signal=f'{DatasetCreado.__name__}Integracion')
