+++ src/mapear/modulos/mapear/aplicacion/comandos/base.py
@@ -1,8 +1,8 @@
-from mapear.seedwork.aplicacion.queries import QueryHandler
-from mapear.modulos.procesamiento.infraestructura.fabricas import FabricaRepositorio
-from mapear.modulos.procesamiento.dominio.fabricas import FabricaMapear
+from mapear.seedwork.aplicacion.comandos import ComandoHandler
+from mapear.modulos.mapear.infraestructura.fabricas import FabricaRepositorio
+from mapear.modulos.mapear.dominio.fabricas import FabricaMapear
 
-class QueryBaseHandler(QueryHandler):
+class CrearBaseHandler(ComandoHandler):
     def __init__(self):
         self._fabrica_repositorio: FabricaRepositorio = FabricaRepositorio()
         self._fabrica_mapear: FabricaMapear = FabricaMapear()
