+++ src/mapear/modulos/mapear/dominio/mixins.py
