+++ src/mapear/modulos/mapear/dominio/objetos_valor.py
