+++ src/mapear/modulos/mapear/dominio/reglas.py
