+++ src/mapear/modulos/mapear/infraestructura/dto.py
