+++ src/mapear/modulos/mapear/infraestructura/schema/v1/comandos.py
