+++ src/mapear/modulos/mapear/infraestructura/schema/v1/eventos.py
